# Conflicts:
#	Modules/Havana/Source/Widgets/AssetBrowser.h
#	Source/Components/Audio/AudioSource.cpp

+++ Modules/Havana/Source/Widgets/AssetBrowser.h
@@ -134,7 +134,7 @@ private:
 	std::vector<AssetDescriptor*> FilteredAssetList;
 	AssetType ForcedAssetFilter = AssetType::Unknown;
 	bool items_need_filtered = true;
-	bool assetTypeFilters[AssetType::Count];
+	bool assetTypeFilters[static_cast<unsigned int>(AssetType::Count)];
 
 	Path AssetBrowserPath;
 	MetaBase* metafile = nullptr;
